+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Initial release with:
 - mode()
 - frequencies(): a frequency is the number of times a value of the data occurs;
 - relativeFrequencies(): a relative frequency is the ratio (fraction or proportion) of the number of times a value of the data occurs in the set of all outcomes to the total number of outcomes;
-- getCumulativeFrequences(): is the accumulation of the previous relative frequencies.;
+- cumulativeFrequencies(): is the accumulation of the previous relative frequencies.;
 - cumulativeRelativeFrequencies(): is the accumulation of the previous relative ratio.
 
 ## 0.1.0 - 2022-01-08
@@ -29,5 +29,5 @@ Initial release with:
 - mode()
 - frequencies(): a frequency is the number of times a value of the data occurs;
 - relativeFrequencies(): a relative frequency is the ratio (fraction or proportion) of the number of times a value of the data occurs in the set of all outcomes to the total number of outcomes;
-- getCumulativeFrequences(): is the accumulation of the previous relative frequencies.;
+- cumulativeFrequencies(): is the accumulation of the previous relative frequencies.;
 - cumulativeRelativeFrequencies(): is the accumulation of the previous relative ratio.